pyproject.toml
[project]
name = "samenwijzer"
version = "0.1.0"
description = "Python app die AI en Data gebruikt om studenten te ondersteunen bij het leren"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.40.0",
    "pandas>=2.2.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "altair>=6.0.0",
    "twilio>=9.10.5",
    "fastapi>=0.135.3",
    "uvicorn>=0.44.0",
    "cryptography>=46.0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
    "ty>=0.0.1a6",
]

[tool.uv]
cache-dir = "./.uv_cache"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
    "ty>=0.0.1a6",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"src/samenwijzer/styles.py" = ["E501"]  # HTML strings zijn inherent lang
"app/main.py" = ["E501"]               # HTML strings zijn inherent lang
"app/pages/*.py" = ["E501"]            # HTML strings zijn inherent lang

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src/samenwijzer --cov-report=term-missing"

[tool.ty]
src = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
